splitting by topic, adding grapher steps
