+++ etl/steps/data/garden/education/2023-07-17/education_lee_lee.py
@@ -10,6 +10,49 @@ from etl.helpers import PathFinder, create_dataset
 
 # Get paths and naming conventions for current step.
 paths = PathFinder(__file__)
+REGIONS = [
+    "North America",
+    "South America",
+    "Europe",
+    "European Union (27)",
+    "Africa",
+    "Asia",
+    "Oceania",
+    "Low-income countries",
+    "Upper-middle-income countries",
+    "Lower-middle-income countries",
+    "High-income countries",
+]
+
+
+def add_data_for_regions(tb: Table, regions: List[str], ds_regions: Dataset, ds_income_groups: Dataset) -> Table:
+    tb_with_regions = tb.copy()
+
+    aggregations = {
+        column: "median"
+        for column in tb_with_regions.columns
+        if column not in ["country", "year"] and "number" not in column
+    }
+
+    for region in REGIONS:
+        # Find members of current region.
+        members = geo.list_members_of_region(
+            region=region,
+            ds_regions=ds_regions,
+            ds_income_groups=ds_income_groups,
+        )
+        tb_with_regions = geo.add_region_aggregates(
+            df=tb_with_regions,
+            region=region,
+            countries_in_region=members,
+            countries_that_must_have_data=[],
+            num_allowed_nans_per_year=None,
+            frac_allowed_nans_per_year=0.99999,
+            aggregations=aggregations,
+        )
+    tb_with_regions = tb_with_regions.copy_metadata(from_table=tb)
+
+    return tb_with_regions
 
 
 def run(dest_dir: str) -> None:
@@ -19,6 +62,11 @@ def run(dest_dir: str) -> None:
     # Load meadow dataset.
     ds_meadow = cast(Dataset, paths.load_dependency("education_lee_lee"))
 
+    # Load regions dataset.
+    ds_regions: Dataset = paths.load_dependency("regions")
+
+    # Load income groups dataset.
+    ds_income_groups: Dataset = paths.load_dependency("income_groups")
     # Read table from meadow dataset.
     tb = ds_meadow["education_lee_lee"]
     tb.reset_index(inplace=True)
@@ -35,6 +83,8 @@ def run(dest_dir: str) -> None:
             "not specified": "Age not specified",
         }
     )
+    tb = add_data_for_regions(tb=tb, regions=REGIONS, ds_regions=ds_regions, ds_income_groups=ds_income_groups)
+
     # Clean and pivot enrollment data by sex
     df_enrollment = prepare_enrollment_data(tb)
     # Clean and pivot attainment data by sex and age
