+++ etl/steps/data/grapher/wb/2023-07-10/education_saber.py
@@ -15,10 +15,10 @@ def run(dest_dir: str) -> None:
     # Load inputs.
     #
     # Load garden dataset.
-    ds_garden = cast(Dataset, paths.load_dependency("education"))
+    ds_garden = cast(Dataset, paths.load_dependency("education_saber"))
 
     # Read table from garden dataset.
-    tb = ds_garden["education"]
+    tb = ds_garden["education_saber"]
 
     #
     # Process data.
